chore: move versions to its own internal file
