+++ internal/common/defaults.go
@@ -5,8 +5,6 @@
 package common
 
 import (
-	"strings"
-
 	"github.com/elastic/e2e-testing/internal/deploy"
 	"github.com/elastic/e2e-testing/internal/shell"
 	"github.com/elastic/e2e-testing/internal/utils"
@@ -80,53 +78,6 @@ func init() {
 			"apm-environment": shell.GetEnv("ELASTIC_APM_ENVIRONMENT", "local"),
 		}).Info("Current execution will be instrumented 🛠")
 	}
-}
-
-// ElasticVersion represents a version
-type elasticVersion struct {
-	Version         string // 8.0.0
-	FullVersion     string // 8.0.0-abcdef-SNAPSHOT
-	HashedVersion   string // 8.0.0-abcdef
-	SnapshotVersion string // 8.0.0-SNAPSHOT
-}
-
-func newElasticVersion(version string) *elasticVersion {
-	versionWithoutCommit := utils.RemoveCommitFromSnapshot(version)
-	versionWithoutSnapshot := strings.ReplaceAll(version, "-SNAPSHOT", "")
-	versionWithoutCommitAndSnapshot := strings.ReplaceAll(versionWithoutCommit, "-SNAPSHOT", "")
-
-	ev := &elasticVersion{
-		FullVersion:     version,
-		HashedVersion:   versionWithoutSnapshot,
-		SnapshotVersion: versionWithoutCommit,
-		Version:         versionWithoutCommitAndSnapshot,
-	}
-
-	if utils.SnapshotHasCommit(version) {
-		ev.HashedVersion = versionWithoutSnapshot
-	}
-
-	return ev
-}
-
-// GetCommitVersion returns a version including the version and the git commit
-func GetCommitVersion(version string) string {
-	return newElasticVersion(version).HashedVersion
-}
-
-// GetFullVersion returns a version including the full version: version, git commit and snapshot
-func GetFullVersion(version string) string {
-	return newElasticVersion(version).FullVersion
-}
-
-// GetSnapshotVersion returns a version including the snapshot version: version and snapshot
-func GetSnapshotVersion(version string) string {
-	return newElasticVersion(version).SnapshotVersion
-}
-
-// GetVersion returns a version without snapshot or commit
-func GetVersion(version string) string {
-	return newElasticVersion(version).Version
 }
 
 // InitVersions initialise default versions. We do not want to do it in the init phase
